Week cards implemented
"Week cards" are playing cards that represent the ISO Week number. Since
there are 53 cards in the reduced French Tarot (including one of the two
Jokers), and there are up to 53 ISO weeks in a year, a one-to-one
correlation can be made between playing cards and week numbers. In this
implementation, this is done by assigning week numbers to a deck in New
Deck Order (Ace of Hearts to King of Hearts, Ace of Clubs to King of
Clubs, King of Diamonds to Ace of Diamonds, Kind of Spades to Ace of
Spades), thus week 1 is the Ace of Hearts, week 2 is the Deuce of
Hearts, etc. If there is a 53rd week that year, the Joker is assigned to
that week.