fix: Last trimester/quarter page shows wrong FY
The last trimester/quarter page showed the incorrect Fiscal Year in the
page title. As a part of fixing this, a significant refactor of the
Planning Calendar tool was undertaken. Functionality appears to remain
the same as before, minus the bug mentioned above.

Additionally, the need to have a Lunar Calibration Date in the config
has been eliminated. The config value can remain (or be removed), but it
will be ignored as of this version.