feat: Added three of the 4 Greater Sabbats
In my opinion, the Greater Sabbats of neo-paganism better describe the
startpoints of the 4 seasons, especially the way we tend to view the
four seasons. So, while I'm not pagan, I decided I wanted to mark them
on the Planning Calendar.

The festival that got left out was Lughnasadh because it coincides with
the Swiss National Holiday. Right now, the calendar doesn't have great
support for multiple holidays landing on the same day, so the code is
present but commented out.